refactor: KG to use the moved finder from the entities-search
