+++ entities-search/src/main/scala/io/renku/entities/search/WorkflowsQuery.scala
@@ -16,8 +16,7 @@
  * limitations under the License.
  */
 
-package io.renku.knowledgegraph.entities
-package finder
+package io.renku.entities.search
 
 import Criteria.Filters.EntityType
 import cats.data.NonEmptyList
@@ -25,8 +24,8 @@ import cats.syntax.all._
 import io.circe.{Decoder, DecodingFailure}
 import io.renku.graph.model.entities.{CompositePlan, StepPlan}
 import io.renku.graph.model.{plans, projects}
-import io.renku.knowledgegraph.entities.model.Entity.Workflow.WorkflowType
-import io.renku.knowledgegraph.entities.model.{Entity, MatchingScore}
+import model.Entity.Workflow.WorkflowType
+import model.{Entity, MatchingScore}
 
 private case object WorkflowsQuery extends EntityQuery[model.Entity.Workflow] {
