+++ knowledge-graph/src/test/scala/io/renku/knowledgegraph/MicroserviceRoutesSpec.scala
@@ -245,9 +245,10 @@ class MicroserviceRoutesSpec
   }
 
   "GET /knowledge-graph/entities" should {
-    import entities.Criteria.Sorting._
-    import entities.Criteria._
-    import entities._
+    import io.renku.entities.search.Criteria
+    import io.renku.entities.search.Criteria.Sorting._
+    import io.renku.entities.search.Criteria._
+    import io.renku.entities.search.Generators._
 
     forAll {
       Table(
@@ -850,7 +851,7 @@ class MicroserviceRoutesSpec
     val projectPathAuthorizer      = mock[Authorizer[IO, model.projects.Path]]
     val datasetIdAuthorizer        = mock[Authorizer[IO, model.datasets.Identifier]]
     val routesMetrics              = TestRoutesMetrics()
-    val versionRoutes              = mock[version.Routes[IO]]
+    private val versionRoutes      = mock[version.Routes[IO]]
 
     def routes(maybeAuthUser: Option[AuthUser] = None): Resource[IO, Kleisli[IO, Request[IO], Response[IO]]] = routes(
       givenAuthIfNeededMiddleware(returning = OptionT.some[IO](maybeAuthUser))
